requirements.txt
gradio>=4.44.1
transformers>=4.41.0
torch>=2.2.0
timm>=0.9.16
accelerate>=0.30.0
Pillow
